requirements.txt
pytubefix
python-telegram-bot
ffmpeg-python